simplify app_info to get info for one app only
